feat: add Bollinger Bands strategy implementation and UI controls
- Implemented Bollinger Bands strategy in the leveraged trend hold module with parameters for period, standard deviation multiplier, trend period, squeeze lookback, squeeze ratio, and ATR stop multiplier.
- Added functionality for updating and checking Bollinger Bands conditions for entries and exits in the trading strategy.
- Created API hooks for managing auto trading budget, including fetching and updating budget data.
- Developed UI components for auto trading budget settings, allowing users to input and manage allocated amounts in KRW and USD.
- Introduced Bollinger control UI for configuring strategy parameters with validation and user feedback.

+++ src-tauri/src/trading/strategy/breakout.rs
@@ -325,8 +325,7 @@ impl Strategy for ConsecutiveMoveStrategy {
         if !self.config.targets_symbol(symbol) {
             return;
         }
-        let cap =
-            bounded_window_with_extra(self.params.buy_days.max(self.params.sell_days) as usize, 1);
+        let cap = bounded_window_with_extra(self.params.buy_days.max(self.params.sell_days), 1);
         self.states
             .entry(symbol.to_string())
             .or_insert_with(|| ConsecutiveMoveState {
@@ -486,7 +485,7 @@ impl Strategy for FailedBreakoutStrategy {
         if !self.config.targets_symbol(symbol) {
             return;
         }
-        let lookback = bounded_window(self.params.lookback_days as usize);
+        let lookback = bounded_window(self.params.lookback_days);
         let state = self
             .states
             .entry(symbol.to_string())
